requirements.txt
Flask
python-dotenv
waitress
requests
psycopg2-binary
flask_wtf
yamllint
ruamel.yaml
playwright